Revert "Done fix hoempage UI"
This reverts commit fc173d5115e6b8654e7f69693c2596ddd12ade57.
