--- lib/gen/fonts.gen.dart
@@ -1,18 +1,0 @@
-/// GENERATED CODE - DO NOT MODIFY BY HAND
-/// *****************************************************
-///  FlutterGen
-/// *****************************************************
-
-// coverage:ignore-file
-// ignore_for_file: type=lint
-// ignore_for_file: directives_ordering,unnecessary_import,implicit_dynamic_list_literal,deprecated_member_use
-
-class FontFamily {
-  FontFamily._();
-
-  /// Font family: Merriweather
-  static const String merriweather = 'Merriweather';
-
-  /// Font family: Mulish
-  static const String mulish = 'Mulish';
-}
