+++ lib/widgets/sectionHeader_item.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
 
-import '../../../gen/fonts.gen.dart';
-
+// Section header for 'Now Showing' and 'Popular'
 class SectionHeader extends StatelessWidget {
   final String title;
-  final VoidCallback? onSeeMore;
+  final VoidCallback onSeeMore;
 
-  const SectionHeader({super.key, required this.title, this.onSeeMore});
+  const SectionHeader({super.key, required this.title, required this.onSeeMore});
 
   @override
   Widget build(BuildContext context) {
@@ -15,10 +14,7 @@ class SectionHeader extends StatelessWidget {
       children: [
         Text(
           title,
-          style: const TextStyle(
-              fontFamily: FontFamily.merriweather,
-              fontSize: 20,
-              fontWeight: FontWeight.w800),
+          style: const TextStyle(fontFamily: 'Merriweather', fontSize: 20, fontWeight: FontWeight.w800),
         ),
         TextButton(
           onPressed: onSeeMore,
@@ -26,11 +22,12 @@ class SectionHeader extends StatelessWidget {
             padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 2),
             decoration: const ShapeDecoration(
                 shape: StadiumBorder(
-              side: BorderSide(color: Colors.grey, width: 1),
-            )),
-            child: const Text('See more',
-                style: TextStyle(
-                    color: Colors.grey, fontFamily: FontFamily.mulish, fontSize: 12)),
+                  side:BorderSide(color: Colors.grey, width: 1),
+                )
+            ),
+            child: const Text(
+                'See more',
+                style: TextStyle(color: Colors.grey,fontFamily: 'Mulish', fontSize:12)),
           ),
         ),
       ],
